refactor proposal create route
