+++ src/app/routers/proposal_create.py
@@ -1,0 +1,33 @@
+from fastapi import APIRouter, Header
+
+from app.contracts.proposals import ProposalCreateEnvelopeResponse, ProposalCreateRequest
+from app.middleware.correlation import correlation_id_var
+from app.services.advisory_service_provider import proposal_service
+
+router = APIRouter(prefix="/api/v1/proposals", tags=["proposals"])
+
+
+@router.post(
+    "",
+    response_model=ProposalCreateEnvelopeResponse,
+    summary="Create Proposal",
+    description=(
+        "Creates a new advisory proposal draft in lotus-advise using a caller-supplied "
+        "idempotency key."
+    ),
+)
+async def create_proposal(
+    request: ProposalCreateRequest,
+    idempotency_key: str = Header(
+        alias="Idempotency-Key",
+        description="Caller-supplied idempotency key for proposal creation requests.",
+        examples=["idem-create-1"],
+    ),
+) -> ProposalCreateEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.create_proposal(
+        body=request.body,
+        idempotency_key=idempotency_key,
+        correlation_id=correlation_id,
+    )
