+++ src/app/routers/proposals.py
@@ -1,8 +1,6 @@
-from fastapi import APIRouter, Header, Path, Query
+from fastapi import APIRouter, Path, Query
 
 from app.contracts.proposals import (
-    ProposalCreateEnvelopeResponse,
-    ProposalCreateRequest,
     ProposalDetailEnvelopeResponse,
     ProposalListEnvelopeResponse,
 )
@@ -10,32 +8,6 @@ from app.middleware.correlation import correlation_id_var
 from app.services.advisory_service_provider import proposal_service
 
 router = APIRouter(prefix="/api/v1/proposals", tags=["proposals"])
-
-
-@router.post(
-    "",
-    response_model=ProposalCreateEnvelopeResponse,
-    summary="Create Proposal",
-    description=(
-        "Creates a new advisory proposal draft in lotus-advise using a caller-supplied "
-        "idempotency key."
-    ),
-)
-async def create_proposal(
-    request: ProposalCreateRequest,
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal creation requests.",
-        examples=["idem-create-1"],
-    ),
-) -> ProposalCreateEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.create_proposal(
-        body=request.body,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
 
 
 @router.get(
